+++ src/features/App/notifications.js
@@ -1,3 +1,12 @@
 // @flow
 
-export default [];
+import T from 'i18n-react';
+
+export default [{
+  id: 'didYouKnow1',
+  message: T.translate('notifications.didYouKnow1'),
+  options: {
+    type: 'info',
+    showOnce: true,
+  },
+}];