+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+
+## [4.26.0][] - 2017-05-21
 ### Changed
 - Webpack to v2
 - Webpack configuration into a factory
@@ -791,7 +793,8 @@ At a later date a embed generator page will be made to help with the creation of
 ### Added
 - This CHANGELOG file to allow more insight to the changes made throughout the development of gw2armory.com
 
-[Unreleased]: https://github.com/madou/armory-react/compare/v4.25.0...HEAD
+[Unreleased]: https://github.com/madou/armory-react/compare/v4.26.0...HEAD
+[4.26.0]: https://github.com/madou/armory-react/compare/v4.25.0...v4.26.0
 [4.25.0]: https://github.com/madou/armory-react/compare/v4.24.0...v4.25.0
 [4.24.0]: https://github.com/madou/armory-react/compare/v4.23.1...v4.24.0
 [4.23.1]: https://github.com/madou/armory-react/compare/v4.23.0...v4.23.1